block: cfq-iosched: Update implementation method
This reverts commit 7271ab97d96fb7b672cbca0

Base Linux Upstream commit:
cfq-iosched: fix the setting of IOPS mode on SSDs

A previous commit wanted to make CFQ default to IOPS mode on
non-rotational storage, however it did so when the queue was
initialized and the non-rotational flag is only set later on
in the probe.

Add an elevator hook that gets called off the add_disk() path,
at that point we know that feature probing has finished, and
we can reliably check for the various flags that drivers can
set.

Fixes: 41c0126 ("block: Make CFQ default to IOPS mode on SSDs")
Tested-by: Romain Francoise <romain@orebokech.com>
Signed-off-by: Jens Axboe <axboe@fb.com>
[Aquarius223: BacKport Linux 3.4.113]
Change-Id: I8e2a310625fcd9be9a9574dceb08215ff2cde6cd